+++ ezcv/content.py
@@ -1,9 +1,8 @@
 """This module is for handling all the functionality around content including:
 - Getting the list of content directories
 - Getting the metadata and contents of content files for sections
+- Classes for parsing all extensions
 """
-
-
 # Standard Lib Dependencies
 import os                                                # Used primarily in path validation
 from collections import defaultdict                      # Used to give dicts default args
@@ -12,8 +11,41 @@ from typing import DefaultDict, List, Tuple, Type, Union # Used to provide accur
 
 
 # Third Party Dependencies
+import exifread            # Used to get metadata of image files
 import markdown            # Used to render and read markdown files
 from colored import fg     # Used to highlight output with colors, especially errors/warnings
+
+#TODO: Update docstrings
+def get_content_directories() -> List[str]:
+    """Gets a list of the existing content directories i.e. ["projects", "education"]
+
+    Returns
+    -------
+    List[str]:
+        The list of existing content directories i.e. ["projects", "education"]
+    """
+    result:list[str] = []
+    for current_path in os.listdir("content"):
+        if os.path.isdir(os.path.join("content", current_path)):
+            result.append(os.path.join("content", current_path))
+    return result
+
+
+def get_section_content(section_folder: str, examples: bool = False) -> List[List[Union[defaultdict, str]]]:
+    content:List[List[Union[defaultdict, str]]] = []
+    extension_handlers:DefaultDict[str, Type] = Content.get_available_extensions()
+
+    for file_name in os.listdir(section_folder):
+        if not examples and file_name.startswith("example"):
+            continue
+        else:
+            extension = "." + file_name.lower().split(".")[-1]
+            if extension_handlers[extension]:
+                extension_handler = extension_handlers[extension]()
+                metadata, html = extension_handler.get_content(os.path.join(section_folder, file_name))
+                content.append([metadata, html])
+    return content
+
 
 @dataclass
 class Content(dict):
@@ -24,6 +56,8 @@ class Content(dict):
     - All subclasses are assumed to have implemented:
         - __metadata__(); Returns a defaultdict of metadata
         - __html__(); Returns the HTML to render
+        - A list attribute called extensions, for example in markdown it would be
+            extensions:List[str] = ['.md', '.markdown', '.mdown', '.mkdn', '.mkd', '.mdwn']
 
     Methods
     -------
@@ -34,6 +68,39 @@ class Content(dict):
     ------
     NotImplementedError
         If any of __metadata__(), or __html__() are not implemented in subclass
+
+    Examples
+    --------
+    ### Get the list of extensions and their handlers
+    ```
+    extension_handlers = Content.get_available_extensions()
+    print(extension_handlers) ''' defaultdict(
+        <function Content.get_available_extensions.<locals>.<lambda> at 0x00000240878690D0>, 
+        {
+        '.md': <class '__main__.Markdown'>, 
+        '.markdown': <class '__main__.Markdown'>,
+        '.jpg': <class '__main__.Image'>,
+        '.png': <class '__main__.Image'>
+        }
+    )'''
+
+    # Get an extension handler for a specific extension (in thise case .md files)
+    print(extension_handlers[".md"]) # <class '__main__.Markdown'>
+    ```
+
+    ### Get content of a list of files
+    ```
+    content = [] # The list that
+    filenames = ['file.md', 'image.jpg'] # Gotten from somewhere else
+    extension_handlers = Content.get_available_extensions()
+
+    for current_file in filenames:
+        extension = '.' + current_file.split('.')[-1]
+        if extension_handlers[extension]:
+            extension_handler = extension_handlers[extension]()
+            metadata, html = extension_handler.get_content(current_file)
+            content.append([metadata, html])
+    ```
     """
 
 
@@ -99,6 +166,18 @@ class Content(dict):
 class Markdown(Content):
     """Used for parsing markdown content
 
+    Notes
+    -----
+    - Default extensions are 
+        - [footnotes](https://python-markdown.github.io/extensions/footnotes/)
+        - [tables](https://python-markdown.github.io/extensions/tables/)
+        - [toc (Table of contents)](https://python-markdown.github.io/extensions/toc/)
+        - [abbr(abbreviations)](https://python-markdown.github.io/extensions/abbreviations/)
+        - [def_list(Definition lists)](https://python-markdown.github.io/extensions/definition_lists/)
+        - [sane_lists(Sane lists)](https://python-markdown.github.io/extensions/sane_lists/)
+        - [meta](https://python-markdown.github.io/extensions/meta)
+        - [mdx_math](https://github.com/mitya57/python-markdown-math)
+
     Examples
     --------
     ```
@@ -107,10 +186,10 @@ class Markdown(Content):
     html, metadata = Markdown().get_content('file_1.md')
     ```
     """
-    md:markdown.Markdown = markdown.Markdown(extensions=['meta']) # Setup markdown parser with extensions
+    md:markdown.Markdown = markdown.Markdown(extensions=['meta', 'footnotes', 'tables', 'toc', 'abbr', 'def_list', 'sane_lists', "mdx_math"]) # Setup markdown parser with extensions
     extensions:List[str] = (".md", ".markdown", ".mdown", ".mkdn", ".mkd", ".mdwn")
 
-
+    #TODO: Add docs about new extensions including adding to example sites, and add latex support to all sites
     def __metadata__(self) -> defaultdict:
         """Gets the metadata from the YAML frontmatter of the markdown file
 
@@ -171,103 +250,77 @@ class Markdown(Content):
             raise FileNotFoundError(f"{fg(1)} Could not find file: {file_path}{fg(15)}\n")
         html = self.__html__(file_path)
         metadata = self.__metadata__()
-        return html, metadata
+        return metadata, html
 
 
-# def get_section_content_new(section_folder: str, examples: bool = False) -> List[Type[Content]]:
-#     content:List[Type[Content]] = []
-
-#     extension_handlers:DefaultDict[str, Type] = Content.get_available_extensions()
-#     if not examples:
-#         filenames = [os.path.abspath(filename) for filename in os.listdir(section_folder) if os.path.isfile(filename) and not filename.startswith("example")]
-#     else:
-#         filenames = [os.path.abspath(filename) for filename in os.listdir(section_folder) if os.path.isfile(filename)]
-#     for current_file in filenames:
-#         extension = "." + current_file.split(".")[-1]
-#         if extension_handlers[extension]:
-#             with extension_handlers[extension](current_file) as file_instance:
-#                 content.append(file_instance)
-#     ...
-
-
-def get_content_directories() -> List[str]:
-    """Gets a list of the existing content directories i.e. ["projects", "education"]
-
-    Returns
-    -------
-    List[str]:
-        The list of existing content directories i.e. ["projects", "education"]
+@dataclass
+class Image(Content):
+    """Used for parsing images
+    
+    Notes
+    -----
+    - Exif data is NOT available on PNG images (it is only available of jpg and tiff)
+    - Since there are so many conditionals it is recommended to use the existing gallery stylesheet
     """
-    result:list[str] = []
-    for current_path in os.listdir("content"):
-        if os.path.isdir(os.path.join("content", current_path)):
-            result.append(os.path.join("content", current_path))
-    return result
+    ignore_exif_data:bool = False
+    extensions:List[str] = (".jpg", ".png", ".tiff", ".avix")
 
 
-# NOTE: This will be deprecated and replaced with the new content system in the next release
-def get_section_content(section_folder: str, examples: bool = False) -> List[List[Union[defaultdict, str]]]:
-    """Gets the markdown content and metadata from each file in a given section
+    def __metadata__(self, filename:str) -> defaultdict:
+        if self.ignore_exif_data:
+            return defaultdict(lambda:False)
 
-    Parameters
-    ----------
-    section_folder : (str)
-        The folder path to the sections content folder i.e. ./content/projects
+        elif filename.lower().endswith("jpg") or filename.lower().endswith("tiff"):
+            with open(filename ,"rb") as f:
+                tags = exifread.process_file(f)
+            tags = defaultdict(lambda:False, tags)
+            return tags
 
-    examples : (bool, optional)
-        Whether or not to include markdown files that have example in the name, by default False
+        else:
+            return defaultdict(lambda:False)
 
-    Returns
-    -------
-    List[List[defaultdict, str]]:
-        A list of lists with the metadata of each page at index 0 and the content at index 1,
-            returns and empty list if the folder does not exist or is empty
 
-    Examples
-    --------
-    Get the contents of the ./projects content folder
-    ```
-    from ezcv.content import get_section_content
+    def __html__(self, tags:defaultdict) -> str:
+        #TODO: Document class names
+        html = ""
 
-    contents = get_section_content("projects")
-    ```
-    """
-    if not os.path.exists(section_folder):
-        return []
+        # Lens detail
+        if tags['EXIF LensModel']:
+            html += f"<p class='lens'>{tags['EXIF LensModel']}</p>\n"
+        
+        # Focal length
+        if tags['EXIF FocalLengthIn35mmFilm']:
+            if tags['EXIF FocalLengthIn35mmFilm'] != tags['EXIF FocalLength']:
+                html += f"<p class='focal-length'>{tags['EXIF FocalLengthIn35mmFilm']}mm (full frame equivalent)</p>\n"
+            else:
+                html += f"<p class='focal-length'>{tags['EXIF FocalLengthIn35mmFilm']}mm</p>\n"
+        else:
+            if tags['EXIF FocalLength']:
+                html += f"<p class='focal-length'>{tags['EXIF FocalLength']}mm</p>\n"
 
-    contents:List[List[defaultdict, str]] = []
+        # ISO, Shutter speed, Apperture
+        if tags['EXIF ISOSpeedRatings']:
+            html += f"<p class='iso'>ISO {tags['EXIF ISOSpeedRatings']}</p>\n"
+        if tags['EXIF ExposureTime']:
+            html += f"<p class='shutter-speed'>{tags['EXIF ExposureTime']} Second(s)</p>\n"
+        if tags['EXIF FNumber']:
+            from fractions import Fraction
+            tags['EXIF FNumber'] = str(float(Fraction(str(tags['EXIF FNumber'])))) # Convert aperture to str i.e. 6.3
+            html += f"<p class='aperture'>f{tags['EXIF FNumber']}</p>\n"
 
-    # Iterate through project files and generate markdown
-    for content in os.listdir(section_folder): 
-        if examples:
-            if content.endswith(".md") or content.endswith(".markdown"):
-                with open(f"{section_folder}{os.sep}{content}", "r") as mdfile: # Parse markdown file
-                    text = mdfile.read()
+        # Camera body details
+        if tags['Image Make'] and tags['Image Model']:
+            html += f"<p class='camera-type'>{tags['Image Make']} {tags['Image Model']}</p>\n"
+        elif tags['Image Make']:
+            html += f"<p class='camera-type'>{tags['Image Make']}</p>\n"
+        elif tags["Image Model"]:
+            html += f"<p class='camera-type'>{tags['Image Model']}</p>\n"
+        else:
+            ...
+        return html
 
-                md = markdown.Markdown(extensions=['meta']) # Setup markdown parser with extensions
-                page_html = md.convert(text) # Convert the markdown content text to hmtl
-                page_meta = defaultdict(lambda:False)
-                # Grab the metadata from the frontmatter of the markdown file
-                for key in md.Meta: # Copy config dict to default_dict_config
-                    if type(md.Meta[key]) == list:
-                        page_meta[key] = md.Meta[key][0]
-                    else:
-                        page_meta[key] = md.Meta[key]
-                contents.append([page_meta, page_html])
 
-        else: # if files with example at the start
-            if content.endswith(".md") and not content.startswith("example"):
-                with open(f"{section_folder}{os.sep}{content}", "r") as mdfile: # Parse markdown file
-                    text = mdfile.read()
-
-                md = markdown.Markdown(extensions=['meta']) # Setup markdown parser with extensions
-                page_html = md.convert(text) # Convert the markdown content text to hmtl
-                page_meta = defaultdict(lambda:False)
-                # Grab the metadata from the frontmatter of the markdown file
-                for key in md.Meta: # Copy config dict to default_dict_config
-                    if type(md.Meta[key]) == list:
-                        page_meta[key] = md.Meta[key][0]
-                    else:
-                        page_meta[key] = md.Meta[key]
-                contents.append([page_meta, page_html])
-    return contents
+    def get_content(self, file_path: str):
+        tags = self.__metadata__(file_path)
+        html = self.__html__(tags)
+        return tags, html